requirements.txt
librosa>=0.7.0
mido==1.2.9
numpy>=1.20.2
scipy>=1.6.2
tensorboard>=2.4.0
#torch>=1.6.0
torchaudio>=0.7.0
